+++ Assets/VRM10/Editor/Components/SpringBone/SpringBoneTreeView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 735a730511f68da4da8d2cf2dcba722e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
